same fix private

+++ app/controllers/groups_controller.rb
@@ -5,39 +5,41 @@ class GroupsController < ApplicationController
   end
 
   def create
+    if current_user && current_user.role && current_user.role.can_edit_schedule 
       @group = Group.new(params[:group])
       if @group.save
-        
-	@group.day.create(:name=>"Понедельник")
-	@group.day.create(:name=>"Вторник")
-	@group.day.create(:name=>"Среда")
-	@group.day.create(:name=>"Четверг")
-	@group.day.create(:name=>"Пятница")
-	
-	@group.day.all.each do|l|
-	  for i in 1..4
-	    l.lesson.create
-	  end  
-	end
-	redirect_to schedule_path
-	
-	
+        @group.day.create(:name=>"Понедельник")
+        @group.day.create(:name=>"Вторник")
+        @group.day.create(:name=>"Среда")
+        @group.day.create(:name=>"Четверг")
+        @group.day.create(:name=>"Пятница")	
+        @group.day.all.each do|l|
+          for i in 1..4
+            l.lesson.create
+          end  
+        end
+      redirect_to schedule_path
       else
         render :action => :new
-  end
-      
-    end
-
-  def destroy
-    
-    @Group = Group.find(params[:id])
-    @Group.day.all.each do|day|
-      day.lesson.all.each do|lesson|
-	lesson.destroy
       end
-      day.destroy
+    else
+      redirect_to root_path
     end
+  end
+
+  def destroy  
+    if current_user && current_user.role && current_user.role.can_edit_schedule 
+      @Group = Group.find(params[:id])
+      @Group.day.all.each do|day|
+        day.lesson.all.each do|lesson|
+  	      lesson.destroy
+        end
+        day.destroy
+      end
     @Group.destroy
     redirect_to schedule_path
+    end
+  else
+    redirect_to root_path
   end
 end
